Making step size adaptive
I added a starting step size and an ending step size. When an
improvement is not achieved using the current step size, the step size
is lowered. This is done until the ending step size is reached.

The iteration also returns a boolean that says whether or not we're done
iterating altogether.