Adding method to compensate for inherent twist
SplineCoaster::compensateTwist() can now detect the inherent twist of a
space curve, including the initial configured twist, and compensate for
that. This now "builds up" the twist, so for example, a nearly planar
figure 8, when twisting in space to form a circle, ends with a twist of
360 degrees, as expected.

The method attempts to prevent sudden jumps in the compensating twist,
but this fails when the initial twist is something like 180 degrees. I
don't even know if we should allow such a case, so for now, I'm leaving
it as is.