Implementing twist penalty
Improved the twist compensation routine.

Global twist is accomodated for in the energy calculation. There is a
twist penalty weight, but optimal values differ wildly for different
tracks. Also, the weight seems to be very binary, where one value will
cause the bending to dominate, and a slightly higher one will cause the
twist penalty to dominate.