Adding AntTweakBar
The header is checked in. The makefile links with AntTweakBar. Either
the library will be available for dynamic linking, or it can be placed
in ./lib/ for static linking. The library is not included in the repo
due for portability reasons (.gitignore also blacklists it).

main.cpp initializes the library now, but no configuration is provided
just yet.