Added PDFLatex support
Previously, due to the embedded graph, I had to generate a postcript
output, then convert to PDF (this was done by LyX, but it was still
annoying). The problem was that graph was in EPS format, which is not
supported by PDFLatex. I converted the graph to PDF, and it works.