Added KBMTorus model
The first steps toward a 2 Klein Bottle Mouth model. The model is based
on two separate b-splines, neither of which is closed. Currently, the
main program hardcodes the usage of the KBMTorus model instead of
allowing the previously usable tracks.

paper/NOTES has been updated to talk more about this model.