Added length variation penalty + refactoring
Now, when two consecutive struts are not the same length, there is a
penalty proportional to how much they differ by. This allows some
element of the previous spring energy, without forcing all struts to
want to stay close to one hardcoded length. This is useful, for example,
when the struts need to get progressively smaller along a path, just as
long as the changes are not drastic from one strut to another.

This also solves an issue with the optimization, where possibly, a local
minimum was being reached when only penalizing bending.

Also added some methods to KBMTorus so not everything is hardcoded in
KBMEnergy.