Implemented varying cross-sectional scaling
Each control point is now associated with its own cross-sectional
scaling factor that allows for varying scaling along the sweep of the
curve.

The optimization also takes this into account and implements a
placeholder penalty for variations in scaling between subsequent points.
This penalty is just a first attempt, and will most likely be changed
later.