Used log ratio for length deviations
Previously, the penalty for the length deviations between two segments
was deviation from 1 of the ratio between the two lengths. However, this
formulation depends on which length is the numerator and which is the
denominator. Now, the log of the ratio is taken, which will be
equivalent up to negatives, and we don't need to introduce the constant
1 because log1 = 0.