Started adding twist calculation/optimization
I'm not sure if the twist calculation is exactly correct, but I'm using
the method of projecting the normals at each point onto the segment
bisector plane, then simply correcting the global twist based on the sum
of the segment-by-segment twists.

The twist is simply distributed evenly for now (because per segment
twist is not penalized, just the sum), but later may be distributed
differently.