Torus tilt is now a optimization parameter
One major (conceptual) change was to no longer normalize the struts.
This was because the end segments were placed in a particular spot
initially, then moved when the struts were normalized. However, this
meant that changing the location of the end segments based *only* on the
torus tilt was difficult as we don't want to renormalize the struts
whenever the torus tilt is changed.