Calculating global twist across both arms
Instead of penalizing the twist independently on each arm, the twist of
one arm is being used to set the starting up vector of the second arm,
and the difference in angle between the ending up vector of the second
arm (calculated using the forward projection) and the starting up vector
of the first arm (a constant) is used as the global twist.

By removing all the inner points on one of the arms, we see that the
twist penalty is being spread entirely on the longer arm, while the
shorter arm is completely straight.