Offset up vectors in display
Now the vectors show up starting at the boundary of the surface instead
of starting at the center of the underlying curve. This makes it easier
to see both which arm a given vector belongs to and it makes it easier
to see a vector belonging to a fat portion of the tube.