Possible fixes + major experimentation
The twist calculations are now reporting the correct values, but it
makes use of the assumption that the left arm is always straight.

In any case, no single set of parameters seems to be perfect for getting
us the correct optimized result we expect, but we can optimize for one
set of parameters, then fix them or de-emphasize them while starting at
the resulting configuration. A bunch of configurations that were used
are included in the code now.